+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 ### Added
 
 - Adds TraceLogger class for structured logging with unique trace IDs, and creates BaseDTO class with serialization, validation, and redaction capabilities.
+- Add module DB with configuration to generate scripts and execute to run project, update entity id as uuid and change controller, services and repository.